+++ test/VideoContextProgressBar.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import VideoContextProgressBar from '../src/PreviewCanvas/VideoContextPreview';
+
+describe('VideoContextProgressBar', () => {
+  it('should render', () => {
+    // tbd
+  });
+});
